+++ app/tfchain/tfchain.chain.js
@@ -361,7 +361,10 @@ export var TFChainNetworkType =  __class__ ('TFChainNetworkType', [NetworkType],
 		}
 		else {
 		}
-		return 1000;
+		if (self.__eq__ (TFChainNetworkType.DEVNET)) {
+			return 10;
+		}
+		return 720;
 	});},
 	get default_explorer_addresses () {return __get__ (this, function (self) {
 		if (arguments.length) {
@@ -671,7 +674,15 @@ export var GoldChainNetworkType =  __class__ ('GoldChainNetworkType', [NetworkTy
 		}
 		else {
 		}
-		return 1000;
+		if (self.__eq__ (GoldChainNetworkType.TESTNET)) {
+			return 720;
+		}
+		if (self.__eq__ (GoldChainNetworkType.DEVNET)) {
+			return 10;
+		}
+		var __except0__ = ValueError ('invalid Goldchain network type {}'.format (self.value));
+		__except0__.__cause__ = null;
+		throw __except0__;
 	});},
 	get default_explorer_addresses () {return __get__ (this, function (self) {
 		if (arguments.length) {
@@ -960,7 +971,10 @@ export var ThreefoldBonusNetworkType =  __class__ ('ThreefoldBonusNetworkType', 
 		}
 		else {
 		}
-		return 1000;
+		if (self.__eq__ (ThreefoldBonusNetworkType.DEVNET)) {
+			return 10;
+		}
+		return 720;
 	});},
 	get default_explorer_addresses () {return __get__ (this, function (self) {
 		if (arguments.length) {
@@ -1246,7 +1260,10 @@ export var FreeFlowTokenNetworkType =  __class__ ('FreeFlowTokenNetworkType', [N
 		}
 		else {
 		}
-		return 1000;
+		if (self.__eq__ (FreeFlowTokenNetworkType.DEVNET)) {
+			return 10;
+		}
+		return 720;
 	});},
 	get default_explorer_addresses () {return __get__ (this, function (self) {
 		if (arguments.length) {
@@ -1311,6 +1328,325 @@ export var FreeFlowTokenNetworkType =  __class__ ('FreeFlowTokenNetworkType', [N
 Object.defineProperty (FreeFlowTokenNetworkType, 'value', property.call (FreeFlowTokenNetworkType, FreeFlowTokenNetworkType._get_value));;
 FreeFlowTokenNetworkType.TESTNET = FreeFlowTokenNetworkType (1);
 FreeFlowTokenNetworkType.DEVNET = FreeFlowTokenNetworkType (2);
+export var EuroChainNetworkType =  __class__ ('EuroChainNetworkType', [NetworkType], {
+	__module__: __name__,
+	get __init__ () {return __get__ (this, function (self, value) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+						case 'value': var value = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (isinstance (value, EuroChainNetworkType)) {
+			var value = value.value;
+		}
+		else if (isinstance (value, str)) {
+			var value = EuroChainNetworkType.from_str (value).value;
+		}
+		if (!(isinstance (value, int))) {
+			var __except0__ = py_TypeError ('EuroChain network type value was expected to be an int, not be of type {}'.format (py_typeof (value)));
+			__except0__.__cause__ = null;
+			throw __except0__;
+		}
+		if (value < 1 || value > 2) {
+			var __except0__ = ValueError ('EuroChain network type out of range: {}'.format (value));
+			__except0__.__cause__ = null;
+			throw __except0__;
+		}
+		self._value = value;
+	});},
+	get _get_value () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		return self._value;
+	});},
+	get __eq__ () {return __get__ (this, function (self, other) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+						case 'other': var other = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (isinstance (other, EuroChainNetworkType)) {
+			return self.value == other.value;
+		}
+		return self.value == other;
+	});},
+	get __ne__ () {return __get__ (this, function (self, other) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+						case 'other': var other = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		return !(self.__eq__ (other));
+	});},
+	get __int__ () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		return self.value;
+	});},
+	get __str__ () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (self.__eq__ (EuroChainNetworkType.TESTNET)) {
+			return 'testnet';
+		}
+		if (self.__eq__ (EuroChainNetworkType.DEVNET)) {
+			return 'devnet';
+		}
+		var __except0__ = ValueError ('invalid EuroChain network type {}'.format (self.value));
+		__except0__.__cause__ = null;
+		throw __except0__;
+	});},
+	get str () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		return self.__str__ ();
+	});},
+	get from_str () {return __getcm__ (this, function (cls, s) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'cls': var cls = __allkwargs0__ [__attrib0__]; break;
+						case 's': var s = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (!(isinstance (s, str))) {
+			var __except0__ = py_TypeError ('can only convert from a string');
+			__except0__.__cause__ = null;
+			throw __except0__;
+		}
+		var s = jsstr.lower (s);
+		if (__in__ (s, tuple (['testnet', 'test']))) {
+			return EuroChainNetworkType.TESTNET;
+		}
+		if (__in__ (s, tuple (['devnet', 'dev']))) {
+			return EuroChainNetworkType.DEVNET;
+		}
+		var __except0__ = ValueError (s + ' is not a supported EuroChain network type str');
+		__except0__.__cause__ = null;
+		throw __except0__;
+	});},
+	get default_network_type () {return __getcm__ (this, function (cls) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'cls': var cls = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		return EuroChainNetworkType.TESTNET;
+	});},
+	get minimum_miner_fee () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (self.__eq__ (EuroChainNetworkType.DEVNET)) {
+			return Currency ('1.0');
+		}
+		return Currency ('0.1');
+	});},
+	get block_creation_fee () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (self.__eq__ (EuroChainNetworkType.DEVNET)) {
+			return Currency ('10.0');
+		}
+		return Currency (0);
+	});},
+	get miner_payout_block_delay () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (self.__eq__ (EuroChainNetworkType.TESTNET)) {
+			return 720;
+		}
+		if (self.__eq__ (EuroChainNetworkType.DEVNET)) {
+			return 10;
+		}
+		var __except0__ = ValueError ('invalid Eurochain network type {}'.format (self.value));
+		__except0__.__cause__ = null;
+		throw __except0__;
+	});},
+	get default_explorer_addresses () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (self.__eq__ (EuroChainNetworkType.TESTNET)) {
+			return ['https://explorer.testnet.eurochain.nbh-digital.com', 'https://explorer2.testnet.eurochain.nbh-digital.com'];
+		}
+		if (self.__eq__ (EuroChainNetworkType.DEVNET)) {
+			return ['http://localhost:2015'];
+		}
+		var __except0__ = ValueError ('invalid Eurochain network type {}'.format (self.value));
+		__except0__.__cause__ = null;
+		throw __except0__;
+	});},
+	get faucet_address () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		if (self.__eq__ (EuroChainNetworkType.TESTNET)) {
+			return 'https://faucet.testnet.eurochain.nbh-digital.com';
+		}
+		if (self.__eq__ (EuroChainNetworkType.DEVNET)) {
+			return 'http://localhost:2016';
+		}
+		return null;
+	});},
+	get chain_type () {return __get__ (this, function (self) {
+		if (arguments.length) {
+			var __ilastarg0__ = arguments.length - 1;
+			if (arguments [__ilastarg0__] && arguments [__ilastarg0__].hasOwnProperty ("__kwargtrans__")) {
+				var __allkwargs0__ = arguments [__ilastarg0__--];
+				for (var __attrib0__ in __allkwargs0__) {
+					switch (__attrib0__) {
+						case 'self': var self = __allkwargs0__ [__attrib0__]; break;
+					}
+				}
+			}
+		}
+		else {
+		}
+		return Type.EUROCHAIN;
+	});}
+});
+Object.defineProperty (EuroChainNetworkType, 'value', property.call (EuroChainNetworkType, EuroChainNetworkType._get_value));;
+EuroChainNetworkType.TESTNET = EuroChainNetworkType (1);
+EuroChainNetworkType.DEVNET = EuroChainNetworkType (2);
 export var Type =  __class__ ('Type', [object], {
 	__module__: __name__,
 	get __init__ () {return __get__ (this, function (self, value) {
@@ -1342,7 +1678,7 @@ export var Type =  __class__ ('Type', [object], {
 			__except0__.__cause__ = null;
 			throw __except0__;
 		}
-		if (value < 1 || value > 4) {
+		if (value < 1 || value > 5) {
 			var __except0__ = ValueError ('chain type out of range: {}'.format (value));
 			__except0__.__cause__ = null;
 			throw __except0__;
@@ -1438,6 +1774,9 @@ export var Type =  __class__ ('Type', [object], {
 		if (self.__eq__ (Type.GOLDCHAIN)) {
 			return 'goldchain';
 		}
+		if (self.__eq__ (Type.EUROCHAIN)) {
+			return 'eurochain';
+		}
 		if (self.__eq__ (Type.TFBCHAIN)) {
 			return 'tfbchain';
 		}
@@ -1491,6 +1830,9 @@ export var Type =  __class__ ('Type', [object], {
 		if (s == 'goldchain') {
 			return Type.GOLDCHAIN;
 		}
+		if (s == 'eurochain') {
+			return Type.EUROCHAIN;
+		}
 		if (s == 'tfbchain') {
 			return Type.TFBCHAIN;
 		}
@@ -1520,6 +1862,9 @@ export var Type =  __class__ ('Type', [object], {
 		}
 		if (self.__eq__ (Type.GOLDCHAIN)) {
 			return GoldChainNetworkType;
+		}
+		if (self.__eq__ (Type.EUROCHAIN)) {
+			return EuroChainNetworkType;
 		}
 		if (self.__eq__ (Type.TFBCHAIN)) {
 			return ThreefoldBonusNetworkType;
@@ -1551,6 +1896,9 @@ export var Type =  __class__ ('Type', [object], {
 		if (self.__eq__ (Type.GOLDCHAIN)) {
 			return 'GFT';
 		}
+		if (self.__eq__ (Type.EUROCHAIN)) {
+			return 'EUR';
+		}
 		if (self.__eq__ (Type.TFBCHAIN)) {
 			return 'TFB';
 		}
@@ -1567,7 +1915,6 @@ Type.TFCHAIN = Type (1);
 Type.GOLDCHAIN = Type (2);
 Type.TFBCHAIN = Type (3);
 Type.FFTCHAIN = Type (4);
-Type._MIN_CHAIN_TYPE = Type.TFCHAIN;
-Type._MAX_CHAIN_TYPE = Type.FFTCHAIN;
+Type.EUROCHAIN = Type (5);
 
 //# sourceMappingURL=tfchain.chain.map